+++ screens/SelectTime.js
@@ -1,14 +1,14 @@
 import { StyleSheet, SafeAreaView, View, Text, TouchableOpacity, Platform, Button, ScrollView } from "react-native";
-import BackButton from "../../components/buttons/BackButton";
+import BackButton from "../components/buttons/BackButton";
 import React, { useState } from 'react';
 
 import Moment from "moment";
 
 import DateTimePicker from '@react-native-community/datetimepicker';
 
-import textStyle from "../../styles/text";
-import colors from "../../styles/colors";
-import StandardButton from "../../components/buttons/StandardButton";
+import textStyle from "../styles/text";
+import colors from "../styles/colors";
+import StandardButton from "../components/buttons/StandardButton";
 
 const SmallButton = ({ title, action }) => (
     <TouchableOpacity style={ styles.smallButton } onPress={action}>
